+++ CLIENT_EXAMPLES/AstroPhysics.py
@@ -1,7 +1,6 @@
-from dispel4py.base import ProducerPE, IterativePE, ConsumerPE,GenericPE
+from dispel4py.base import IterativePE, GenericPE
 from dispel4py.workflow_graph import WorkflowGraph
-from client import d4pClient,Process
-
+from laminar.client.d4pyclient import d4pClient
 
 type_dict={"E":-5, "E-S0":-3, "S0":-2, "S0a":0, "Sa":1, "Sab":2, "Sb":3, "Sbc":4, "Sc":6, "Scd":7, "Sd":8, "Sm":9, "SBa":1, "SBab":2, "SBb":3, "SBbc":4, "SBc":6, "SBcd":7, "SBd":8, "SBm":9}
