+++ laminar/argument_parser.py
@@ -1,0 +1,13 @@
+import argparse
+import sys
+
+
+class CustomArgumentParser(argparse.ArgumentParser):
+    def __init__(self, exit_on_error=True):
+        super().__init__(exit_on_error=exit_on_error)
+
+    def exit(self, status=0, message=None):
+        if self.exit_on_error:
+            sys.exit(status)
+        else:
+            raise argparse.ArgumentError(None, message=message)
