+++ laminar/global_variables.py
