+++ requirements_client.txt
@@ -16,3 +16,4 @@ accelerate==1.12.0
 antlr4-python3-runtime==4.13.2
 ipython==9.9.0
 findimports==3.0.0
+tabulate==0.9.0
